299: Break tests into smaller chunks
- reorganize utils into more readable file sizes
- reorg directory structure to make navigating a bit easier
