+++ cypress/e2e/Home/home.spec.cy.ts
@@ -1,4 +1,4 @@
-import { setup } from "../support/utils";
+import { setup } from "../../support/setupUtils";
 
 describe("Home", () => {
   before(() => {
@@ -8,7 +8,6 @@ describe("Home", () => {
   it("the UI is available", () => {
     cy.visit("/");
 
-    cy.get("h1")
-      .should("contain", "Welcome To Sundeck OpsCenter");
+    cy.get("h1").should("contain", "Welcome To Sundeck OpsCenter");
   });
 });
