+++ cypress/support/setupUtils.ts
@@ -1,0 +1,3 @@
+export function setup() {
+  Cypress.config("baseUrl", Cypress.env("OPSCENTER_URL"));
+}
